Update tests to mirror codebase
