+++ tests/common/_os/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (C) 2012 Anaconda, Inc
+# SPDX-License-Identifier: BSD-3-Clause
